+++ src/main/java/com/be08/smart_notes/repository/NoteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.be08.smart_notes.entity.NoteEntity;
 
 @Repository
-public interface NoteRepository extends JpaRepository<NoteEntity, Long>{
+public interface NoteRepository extends JpaRepository<NoteEntity, Integer>{
 
 }
